change from javascript to typescript
