+++ App.tsx
@@ -1,4 +1,5 @@
 
+import React from "react";
 import SignedInStack from "./Navigation";
 
 export default function App() {
