+++ components/new post/AddNewPost.tsx
@@ -1,28 +1,25 @@
 import { View, Text, StyleSheet, TouchableOpacity } from "react-native";
 import React from "react";
 import { Ionicons } from "@expo/vector-icons";
-import { useNavigation } from "@react-navigation/native"; 
 import FormikPostUploader from "./FormikPostUploader";
 
-const AddNewPost = () => {
-  const navigation = useNavigation(); 
+const AddNewPost = ({navigation}) => {
   return (
     <View style={styles.container}>
-      <Header />
+      <Header navigation={navigation}/>
       <FormikPostUploader />
     </View>
   );
 };
 
-const Header = () => {
-  const navigation = useNavigation(); 
+const Header = ({ navigation }) => {
   return (
     <View style={styles.headerContainer}>
       <TouchableOpacity onPress={() => navigation.goBack()}>
         <Ionicons name="chevron-back" size={30} color="white" />
       </TouchableOpacity>
       <Text style={styles.headerText}>NEW POST</Text>
-      <View style={{ width: 30 }} /> 
+      <View style={{ width: 30 }} />
     </View>
   );
 };
