+++ package.json
@@ -4,6 +4,7 @@
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
     "@react-navigation/stack": "^6.3.29",
+    "email-validator": "^2.0.4",
     "expo": "^51.0.0",
     "formik": "^2.4.6",
     "react": "18.2.0",
